Simplify module structure and introduce alleycats-retry
