+++ README.md
@@ -2,8 +2,7 @@
 
 A library for retrying actions that can fail.
 
-Designed to work with [cats](https://typelevel.org/cats/) and (optionally)
-[cats-effect](https://typelevel.org/cats-effect/) or [Monix](https://monix.io/).
+Designed to work with [cats](https://typelevel.org/cats/) and [cats-effect](https://typelevel.org/cats-effect/) or [Monix](https://monix.io/).
 
 Inspired by the [retry Haskell
 package](https://hackage.haskell.org/package/retry).
