+++ modules/docs/src/main/mdoc/docs/index.md
@@ -98,8 +98,6 @@ import cats.effect.Timer
 import scala.concurrent.ExecutionContext.global
 implicit val timer: Timer[IO] = IO.timer(global)
 
-// This is so we can use that Timer to perform delays between retries
-import retry.CatsEffect._
 
 val flakyRequestWithRetry: IO[String] =
   retryingOnAllErrors[String](
